+++ pkg/nsx/services/ipaddressallocation/ipaddressallocation.go
@@ -1,7 +1,6 @@
 package ipaddressallocation
 
 import (
-	"context"
 	"fmt"
 	"sync"
